+++ kpops/__init__.py
@@ -1,7 +1,7 @@
 __version__ = "5.0.1"
 
 # export public API functions
-from kpops.api import clean, deploy, destroy, generate, init, manifest, reset
+from kpops.api.api import clean, deploy, destroy, generate, init, manifest, reset
 
 __all__ = (
     "generate",
