+++ kpops/api/exception.py
@@ -7,3 +7,7 @@ class ValidationError(Exception):
 
 class ParsingException(Exception):
     pass
+
+
+class ClassNotFoundError(Exception):
+    """Similar to builtin `ModuleNotFoundError`; class doesn't exist inside module."""
