--- kpops/cli/exception.py
@@ -1,2 +1,0 @@
-class ClassNotFoundError(Exception):
-    """Similar to builtin `ModuleNotFoundError`; class doesn't exist inside module."""
